[fix][refactor][feature] push, pull: figure out upstream branch
- New class `GitBranchPushOrPullBase`, which is a common superclass to
`GitBranchPushed` and `GitBranchPulled`

- Refactor `_deconstruct_remote_name` & co into
`GitBranchPushOrPullBase` and make it private; this is where the
parsing of origin/foo now happens

- `GitBranchPushOrPullBase` will now always set up `.remote` and
`.remote_branch` directly in the constructor, i.e. no more `@{u}`

- `GitSubaction()._to_command_dict`: use a shell if one is
needed (i.e. one arg with a space); in that case, substitute `git`
everywhere (under the assumption that the shell command may invoke git
in a subcommand)